Big I/O rework.
- Properly add priority support. This is accomplished by rearranging the code
  so that the bio module really does all of the i/o and allocators are just
  responsible for making storage available.
- Add uuid support to make identifying devices reliable even if probing is
  inconsistent. This also puts us one step closer to being able to get TuxOnIce
  going on Debian via a simple apt-get install and reboot.

+++ kernel/power/tuxonice_prepare_image.h
@@ -12,7 +12,7 @@
 extern int toi_prepare_image(void);
 extern void toi_recalculate_image_contents(int storage_available);
 extern unsigned long real_nr_free_pages(unsigned long zone_idx_mask);
-extern int image_size_limit;
+extern long image_size_limit;
 extern void toi_free_extra_pagedir_memory(void);
 extern unsigned long extra_pd1_pages_allowance;
 extern void free_attention_list(void);
